requirements.txt
fastapi==0.95.0
pydantic==1.10.7
uvicorn==0.22.0
python-dotenv==1.0.0
transformers==4.28.1
flake8==6.0.0
black==23.3.0